refactoring and file management
