+++ Flutter Project/pubspec.yaml
@@ -38,7 +38,6 @@ dependencies:
   google_fonts: ^4.0.3
   font_awesome_flutter: ^10.4.0
   flutter_glow: ^0.3.0
-  sleek_circular_slider: ^2.0.1
 
 dev_dependencies:
   flutter_test:
